pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-fitness-coach"
version = "1.0.0"
description = "AI-powered fitness coaching environment with workout tracking, BMI calculation, and personalized recommendations"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AI Fitness Team", email = "fitness@example.com"}
]
keywords = ["fitness", "health", "ai", "coaching", "streamlit"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "streamlit>=1.56.0",
    "pandas>=1.3.0",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "openai>=2.30.0",
    "requests>=2.25.0",
    "openenv-core>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
]

[project.urls]
Homepage = "https://github.com/example/ai-fitness-coach"
Repository = "https://github.com/example/ai-fitness-coach.git"
Documentation = "https://github.com/example/ai-fitness-coach/blob/main/README.md"
"Bug Tracker" = "https://github.com/example/ai-fitness-coach/issues"

[project.scripts]
ai-fitness-coach = "app.streamlit_app:main"
server = "server.app:main"

[tool.setuptools]
packages = ["app", "env"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.md"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=app --cov=env --cov-report=html --cov-report=term-missing"
